restructuring packages for prescribe treatment
