+++ src/main/java/usecases/UseCaseValidator.java
@@ -1,0 +1,9 @@
+package usecases;
+
+import java.util.List;
+
+public interface UseCaseValidator {
+    void validate(UseCaseRequest request);
+    boolean isValid();
+    List<String> errors();
+}
