+++ src/main/java/usecases/prescribetreatment/PrescribeTreatmentException.java
@@ -1,0 +1,12 @@
+package usecases.prescribetreatment;
+
+public class PrescribeTreatmentException extends RuntimeException {
+
+    public PrescribeTreatmentException() {
+        super();
+    }
+
+    public PrescribeTreatmentException(String message) {
+        super(message);
+    }
+}
